+++ scripts/cli-migrations/v3/test/seeds/default/1727111758071_init.sql
@@ -1,0 +1,1 @@
+INSERT INTO "public"."test" (id) VALUES (1);